Updated tests for Validators contracts
